Fix "Regolamento didattico" for Maths classes (#167)
* fix: "regolamento" callbacks for math degree

The fact of this issue is the name of the dictionary keys. Python
increasingly reads if-elif-else statements so, when it finds a good key
to the "Regolamento didattico 2020/2021" it does not matter if that "regolamento"
is for class L31 or L35. It returns `True` for the first good `if`.(Same
for LM18 and LM40).

So, it's better define different keys for a different class. Dictionary
keys now end with `_L{class}` so Python can't be wrong!

* fix(tests): press the 4 rules buttons

After pressing the button to obtain a "Regolamento didattico" you get
the "Regolamento didattico" of the current academic year.